Free-Mentors-Backend/requirements.txt
asgiref==3.8.1
Django==3.2.25
django-cors-headers==4.5.0
django-environ==0.12.0
django-filter==23.5
django-graphql-auth==0.3.7
django-graphql-jwt==0.4.0
dnspython==2.7.0
graphene==3.4.3
graphene-django==3.2.2
graphql-core==3.2.6
graphql-relay==3.2.0
promise==2.3
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
six==1.17.0
sqlparse==0.2.4
text-unidecode==1.3
typing_extensions==4.12.2
mongoengine==0.29.0
pymongo==3.12.3
bcrypt==4.1.2 
graphene-mongo==0.4.4